+++ src/main/java/com/biddy/biddy_api/domain/auction/controller/BidController.java
@@ -1,11 +1,10 @@
-package com.biddy.biddy_api.domain.auction;
+package com.biddy.biddy_api.domain.auction.controller;
 
+import com.biddy.biddy_api.domain.auction.service.BidService;
 import com.biddy.biddy_api.domain.auction.dto.BidDto;
 import com.biddy.biddy_api.domain.auction.dto.PostBidRequest;
-import com.biddy.biddy_api.domain.auction.entity.Bid;
 import com.biddy.biddy_api.global.RspTemplate;
 import io.swagger.v3.oas.annotations.Operation;
-import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
