+++ src/main/java/com/biddy/biddy_api/domain/auction/repository/BidRepository.java
@@ -1,4 +1,4 @@
-package com.biddy.biddy_api.domain.auction;
+package com.biddy.biddy_api.domain.auction.repository;
 
 import com.biddy.biddy_api.domain.auction.entity.Bid;
 import org.springframework.data.jpa.repository.JpaRepository;
